requirements.txt
# Project dependencies

python-wordpress-xmlrpc>=2.3.0
requests>=2.31.0
Pillow>=10.1.0
google-api-python-client>=2.108.0
pandas>=2.1.3
openai>=1.3.5
python-dotenv>=1.0.0
pytest>=7.4.3
ratelimit>=2.2.1
configparser>=6.0.0
